refactor: 결제 알림 관련 상수 관리를 위한 PaymentReminderContract 도입
