+++ app/src/main/java/woowacourse/shopping/ui/util/PaymentReminderContract.kt
@@ -1,0 +1,11 @@
+package woowacourse.shopping.ui.util
+
+object PaymentReminderContract {
+    const val ACTION_PAYMENT_REMINDER = "ACTION_PAYMENT_REMINDER"
+
+    const val ACTION_OPEN_PAYMENT_FROM_NOTIFICATION = "OPEN_PAYMENT_FROM_NOTIFICATION"
+
+    const val EXTRA_SELECTED_ITEM_IDS = "extra_selected_item_ids"
+
+    const val EXTRA_NAVIGATE_TO_PAYMENT = "navigate_to_payment"
+}
